Added attack animation on x key
Code is a bit of a mess but works thanks to @BangPeter. I cleaned up some of the earlier code that's unneeded now. A lot of the visual bugs came from the AnimationPlayer Reset track not resetting postion to 0,0 and instead to 20,0.

The attack animation is peck_attack in animationplayer

The attacking and parrying boolean is awkwardly implemented in movement.gd.

Also note that a lot of the AnimationPlayer animations don't move the CollisionShape2D Peck hitbox yet and that should be implemented.

The chicken audio continues to haunt me. please edit that down to the first 2 seconds so you dont get the full 30 second clip. please.

The eventual goal is is one demo level of you vs the farmer. You have 3 HP, he has 3 HP. The farmer walks towards you and uses his stab attack when close.